merge conflicts defined below. SpriteInfo CenterPosition.Z now works in the DX pipeline. Sprites are correctly drawn with its z.

Merge branch 'master' into s2t13_DrawingSpriteThroughPipeline

Conflicts:
	CoreLib/src/Game.cpp
	CoreLib/src/Tilemap.cpp
 It looks like you may be committing a MERGE.

	modified:   CoreLib/src/Game.h
	modified:   CoreLib/src/GameObject.cpp
	modified:   CoreLib/src/GameObject.h
	new file:   TestBacklog/Maps/test_map.txt
	modified:   TestBacklog/TestBacklog.vcxproj
	modified:   TestBacklog/TestBacklog.vcxproj.filters
	modified:   WinLib/src/DxSpriteRenderer.cpp
	modified:   WinLib/src/DxSpriteRenderer.h

+++ LinLib/src/GlSprite.cpp
@@ -64,7 +64,7 @@ int GlSprite::draw()
 		m_spriteInfo.transformInfo.scale[TransformInfo::Y] / 2
 	);
 	glUniform2f(m_spriteShader->getScreenSizeConstant(), 
-		m_context->getScreenWidth(), m_context->getScreenHeight());
+		(GLfloat)m_context->getScreenWidth(), (GLfloat)m_context->getScreenHeight());
 	glActiveTexture(GL_TEXTURE0);
 	glBindTexture(GL_TEXTURE_2D, m_texture);
